requirements.txt
attrs==19.3.0
dateparser==1.0.0
decorator==4.4.2
defusedxml==0.6.0
Django==3.1.3
django-allauth==0.44.0
django-cleanup==5.2.0
django-crispy-forms==1.11.2
django-filter==2.4.0
django-jet==1.0.8
django-rest-auth==0.9.5
djangorestframework==3.12.2
djangorestframework-simplejwt==4.7.1
future==0.18.2
glob2==0.7
greenlet==0.4.16
h5py==2.10.0
html5lib 
imageio 
imagesize==1.2.0
jet-bridge-base==1.0.1
jet-django==1.1.4
Jinja2==2.11.2
joblib 
json5==0.9.5
jsonschema==3.2.0
lxml 
Markdown
numpy==1.18.5 
oauthlib==3.1.0
pandas
path==13.1.0
pathlib2==2.3.5
Pillow 
PyJWT==2.1.0
pyparsing==2.4.7
pyreadline==2.1
pyrsistent==0.16.0
PySocks==1.7.1
pytest==5.4.3
scikit-image==0.16.2
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.15.0
sqlparse==0.4.1
psycopg2==2.7.3.2
testpath==0.4.4
textblob==0.15.3
toolz==0.10.0
tzlocal==2.1
gunicorn==20.0.4
ujson==1.35
urllib3==1.25.9